Add JDBC action classes and refactor query packages
Introduced JdbcExecuteAction and JdbcQueryAction classes with their contexts. Refactored package names from 'on_jdbc_query' to 'on_query' for conditions, triggers, and related tests to improve consistency and clarity.
