+++ automation-engine-jdbc/src/main/java/com/davidrandoll/automation/engine/jdbc/actions/execute/JdbcExecuteActionContext.java
@@ -1,0 +1,4 @@
+package com.davidrandoll.automation.engine.jdbc.actions.execute;
+
+public class JdbcExecuteActionContext {
+}
