+++ automation-engine-jdbc/src/main/java/com/davidrandoll/automation/engine/jdbc/actions/query/JdbcQueryAction.java
@@ -1,0 +1,4 @@
+package com.davidrandoll.automation.engine.jdbc.actions.query;
+
+public class JdbcQueryAction {
+}
