+++ automation-engine-jdbc/src/main/java/com/davidrandoll/automation/engine/jdbc/conditions/on_query/OnJdbcQueryConditionContext.java
@@ -1,4 +1,4 @@
-package com.davidrandoll.automation.engine.jdbc.conditions.on_jdbc_query;
+package com.davidrandoll.automation.engine.jdbc.conditions.on_query;
 
 import com.davidrandoll.automation.engine.core.conditions.IConditionContext;
 import com.fasterxml.jackson.annotation.JsonAlias;
