+++ automation-engine-jdbc/src/main/java/com/davidrandoll/automation/engine/jdbc/triggers/on_query/OnJdbcQueryTrigger.java
@@ -1,7 +1,7 @@
-package com.davidrandoll.automation.engine.jdbc.conditions.on_jdbc_query;
+package com.davidrandoll.automation.engine.jdbc.triggers.on_query;
 
 import com.davidrandoll.automation.engine.core.events.EventContext;
-import com.davidrandoll.automation.engine.spi.PluggableCondition;
+import com.davidrandoll.automation.engine.spi.PluggableTrigger;
 import com.davidrandoll.automation.engine.templating.TemplateProcessor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
@@ -13,35 +13,35 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-@Component("onJdbcQueryCondition")
+@Component("onJdbcQueryTrigger")
 @RequiredArgsConstructor
-@ConditionalOnMissingBean(name = "onJdbcQueryCondition", ignored = OnJdbcQueryCondition.class)
+@ConditionalOnMissingBean(name = "onJdbcQueryTrigger", ignored = OnJdbcQueryTrigger.class)
 @ConditionalOnClass(NamedParameterJdbcTemplate.class)
-public class OnJdbcQueryCondition extends PluggableCondition<OnJdbcQueryConditionContext> {
+public class OnJdbcQueryTrigger extends PluggableTrigger<OnJdbcQueryTriggerContext> {
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final TemplateProcessor pebble;
 
     @Override
-    public boolean autoEvaluateExpression() {
-        return false;
-    }
-
-    @Override
-    public boolean isSatisfied(EventContext ec, OnJdbcQueryConditionContext cc) {
-        if (cc.getQuery() == null || cc.getExpression() == null) {
+    public boolean isTriggered(EventContext ec, OnJdbcQueryTriggerContext tc) {
+        if (tc.getQuery() == null || tc.getExpression() == null) {
             return false;
         }
 
-        List<Map<String, Object>> queryResult = jdbcTemplate.queryForList(cc.getQuery(), cc.getParams());
+        List<Map<String, Object>> queryResult = jdbcTemplate.queryForList(tc.getQuery(), tc.getParams());
         Map<String, Object> context = new HashMap<>();
         context.put("result", queryResult);
         context.put("event", ec);
 
         try {
-            String result = pebble.process(cc.getExpression(), context);
+            String result = pebble.process(tc.getExpression(), context);
             return Boolean.TRUE.toString().equalsIgnoreCase(result);
         } catch (Exception e) {
-            throw new RuntimeException("Error processing expression: " + cc.getExpression(), e);
+            throw new RuntimeException("Error processing expression: " + tc.getExpression(), e);
         }
+    }
+
+    @Override
+    public boolean autoEvaluateExpression() {
+        return false;
     }
 }
