+++ automation-engine-jdbc/src/main/java/com/davidrandoll/automation/engine/jdbc/triggers/on_query/OnJdbcQueryTriggerContext.java
@@ -1,4 +1,4 @@
-package com.davidrandoll.automation.engine.jdbc.triggers.on_jdbc_query;
+package com.davidrandoll.automation.engine.jdbc.triggers.on_query;
 
 import com.davidrandoll.automation.engine.core.triggers.ITriggerContext;
 import com.fasterxml.jackson.annotation.JsonAlias;
